# src/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cadastro",
    "version": "v1"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json; charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/userToRegister"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userToRegister"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/userToRegister"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/userToRegister"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json; charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/userToAuth"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userToAuth"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/userToAuth"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/userToAuth"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Clientes": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": false
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": false
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cliente"
        ],
        "requestBody": {
          "content": {
            "application/json; charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/cliente"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cliente"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/cliente"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/cliente"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Cliente"
        ],
        "requestBody": {
          "content": {
            "application/json; charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/cliente"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cliente"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/cliente"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/cliente"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Clientes/{id}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/cliente"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cliente"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/cliente"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Contato": {
      "post": {
        "tags": [
          "Contato"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contato"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Contato"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Contato"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "userToRegister": {
        "required": [
          "name",
          "email",
          "passwaord"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "passwaord": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "userToAuth": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 8,
            "minLength": 4,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "cliente": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 3,
            "type": "string",
            "nullable": true
          },
          "cidade": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string",
            "nullable": true
          },
          "email": {
            "maxLength": 80,
            "minLength": 3,
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "sexo": {
            "maxLength": 10,
            "minLength": 3,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Contato": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "assunto": {
            "type": "string",
            "nullable": true
          },
          "mensagem": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}